Split the single ecs file into separate ones for the different components
