+++ CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/model/ecs/IEntity.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8237d9b87619e4446b08f04467f2d8f5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
