+++ CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/model/ecs/IEntityExtensions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cfd188c3f7fdc1a4d8ccdaaec9259269
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
